% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_functions.R
\name{matrixification}
\alias{matrixification}
\title{Creates a matrix out of a long list of loadings (or more generally, any list from the psych package)}
\usage{
matrixification(FitData, loadings, factors, cutpoint)
}
\arguments{
\item{FitData}{Data frame from statistical test.}

\item{loadings}{List element of data frame.}

\item{factors}{Number of factors.}

\item{cutpoint}{Value with which you would like data below which to be neglected. Taken as absolute value.}
}
\value{
Matrix to be used as table.
}
\description{
Creates a matrix out of a long list of loadings (or more generally, any list from the psych package)
}
\examples{
PreFitTable7=matrixification(PreFit7,"loadings",c("Pre1","Pre2","Pre3","Pre4","Pre5","Pre6","Pre7"),.3)
}
