% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_functions.R
\name{skreeplot}
\alias{skreeplot}
\title{Creates Skree Plots and returns them as plot for later use.}
\usage{
skreeplot(data, title)
}
\arguments{
\item{data}{Data frame containing only the questions to be analyzed. All data must be mean-imputed.}

\item{title}{Custom titles.}
}
\value{
Actual saved plot of skree plots.
}
\description{
Creates Skree Plots and returns them as plot for later use.
}
\examples{
PreScree=skreeplot(PreDUPI_I,"Skree Plot for Pre-Test \\nRespondents")
}
