% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_functions.R
\name{calcpchisq}
\alias{calcpchisq}
\title{Creates a list of values relating to Chi-Sq Tests.}
\usage{
calcpchisq(x1, x2, df1, df2, lower.tail)
}
\arguments{
\item{x1}{Value for first fit}

\item{x2}{Value for second fit}

\item{df1}{Degrees of freedom for first fit}

\item{df2}{Degrees of freedom for second fit}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X<=x], otherwise, P[X>x].}
}
\value{
List consisting of: differince in fit; difference in degrees of freedom; p-value of Chi-Sq test
}
\description{
Creates a list of values relating to Chi-Sq Tests.
}
\examples{
PreviousChiSq=with(PreviousFit,calcpchisq(null.model,objective,null.dof,dof))
}
